Do not modify assets on uploads
