+++ src/Resources/config/routes_no_locale.yaml
@@ -2,16 +2,13 @@
 # localized URLs (see https://docs.sylius.com/en/latest/cookbook/shop/disabling-localised-urls.html).
 # To accommodate this in your plugin, provide a routes file for non localized stores and tell this in the README.md
 
-setono_sylius_cms_shop:
-    resource: "@SetonoSyliusCMSPlugin/Resources/config/routes/shop.yaml"
-
 setono_sylius_cms_admin:
     resource: "@SetonoSyliusCMSPlugin/Resources/config/routes/admin.yaml"
     prefix: /admin/cms
 
-# todo could this be moved to shop.yaml?
-setono_sylius_cms_download_asset:
-    path: /cms/asset/{id}
-    controller: setono_sylius_cms.controller.action.download_asset
-    requirements:
-        id: \d+
+setono_sylius_cms_global:
+    resource: "@SetonoSyliusCMSPlugin/Resources/config/routes/global.yaml"
+    prefix: /cms
+
+setono_sylius_cms_shop:
+    resource: "@SetonoSyliusCMSPlugin/Resources/config/routes/shop.yaml"
