+++ src/Resources/views/admin/asset/_form.html.twig
@@ -11,17 +11,8 @@
             {{ sscms_preview(resource) }}
 
             <div class="ui segment">
-                {% set assetUrl = url('setono_sylius_cms_download_asset', {'id': resource.id}) %}
-                <div>
-                    URL:
-                    <a href="{{ assetUrl }}" target="_blank">
-                        {{ assetUrl }}
-                    </a>
-                </div>
-                <div>
-                    Path:
-                    {{ resource.path }}
-                </div>
+                <div><a href="{{ url('setono_sylius_cms_view_asset', {'id': resource.id}) }}" target="_blank">View</a></div>
+                <div><a href="{{ url('setono_sylius_cms_download_asset', {'id': resource.id}) }}" target="_blank">Download</a></div>
             </div>
         {% endif %}
     </div>
